+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/config/CloudFrontConfig.java
@@ -1,4 +1,4 @@
-package com.startingblue.fourtooncookie.aws.cloudfront;
+package com.startingblue.fourtooncookie.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
